Check for status = 1 on carousel
Currently, if the published date is null, the carousel will still display, so it’s causing confusion when the editors check the preview mode of the content and see the carousel.

This way, the content has to be published for the carousel to display.  (Draft & Deleted won’t display it in preview)