update auto open logic
Check the location of the link relative to the page to vet if it is left or right and auto apply left right flyout based on which half of the page you are on.  Also make sure there is enough space left or right to open, if not open below.